finished task 3 of Amazing Numbers protject

+++ Amazing Numbers/task/src/numbers/OddDetector.java
@@ -7,7 +7,7 @@ public class OddDetector{
         isOdd = false;
     }
 
-    public boolean evenOrOdd(int num){
+    public boolean evenOrOdd(long num){
         return num % 2 != 0;
     }
 }
